requirements.txt
python-binance==1.0.16
pandas==1.4.3